ocr-meg-collection/pyproject.toml
[tool.poetry]
name = "ocr-meg-collection"
version = "0.1.0"
description = ""
authors = ["LucaZoss <lucazosso@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.43.0"
pandas = "^2.2.2"
requests = "^2.32.3"
pillow = "^10.4.0"
pytesseract = "^0.3.13"
opencv-python = "^4.10.0.84"
matplotlib = "^3.9.2"
ollama = "^0.3.2"
transformers = "^4.44.2"
mutagen = "^1.47.0"
ipykernel = "^6.29.5"
google-cloud-vision = "^3.7.4"
easyocr = "^1.7.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
